+++ Sharpify.CommandLineInterface/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## Version 1.0.3
+
+* Updated `Sharpify` dependency and implemented usage of new APIs to aid in maintainability.
+* Add `DoNotIncludeMetadataInHelpText()` in `CliBuilder` which removes the metadata inclusion in the general help text.
+
 ## Version 1.0.2
 
 * Removed thread-local `StringBuilder` from `CliRunner`, replaced all usages with `StringBuffer` from `Sharpify`